feat(search): move levenshteinDistanceSearch to service
